+++ sdk/src/wallet/storage/participation.rs
@@ -110,3 +110,63 @@ impl StorageManager {
             .unwrap_or_default())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+    use crate::{types::block::payload::transaction::TransactionId, wallet::storage::adapter::memory::Memory};
+
+    #[tokio::test]
+    async fn insert_get_remove_participation_event() {
+        let mut storage_manager = StorageManager::new(Memory::default(), None).await.unwrap();
+        assert!(storage_manager.get_participation_events(0).await.unwrap().is_empty());
+
+        let event_with_nodes = ParticipationEventWithNodes::mock();
+        let event_with_nodes_id = event_with_nodes.id;
+
+        storage_manager
+            .insert_participation_event(0, event_with_nodes.clone())
+            .await
+            .unwrap();
+        let participation_events = storage_manager.get_participation_events(0).await.unwrap();
+
+        let mut expected = HashMap::new();
+        expected.insert(event_with_nodes_id, event_with_nodes);
+
+        assert_eq!(participation_events, expected);
+
+        storage_manager
+            .remove_participation_event(0, &event_with_nodes_id)
+            .await
+            .unwrap();
+        assert!(storage_manager.get_participation_events(0).await.unwrap().is_empty());
+    }
+
+    #[tokio::test]
+    async fn set_get_cached_participation_output_status() {
+        let mut storage_manager = StorageManager::new(Memory::default(), None).await.unwrap();
+        assert!(
+            storage_manager
+                .get_cached_participation_output_status(0)
+                .await
+                .unwrap()
+                .is_empty()
+        );
+
+        let outputs_participation = std::iter::once((
+            OutputId::new(TransactionId::new([3; 32]), 0).unwrap(),
+            OutputStatusResponse::mock(),
+        ))
+        .collect::<HashMap<_, _>>();
+
+        storage_manager
+            .set_cached_participation_output_status(0, outputs_participation.clone())
+            .await
+            .unwrap();
+
+        assert_eq!(
+            storage_manager.get_cached_participation_output_status(0).await.unwrap(),
+            outputs_participation
+        );
+    }
+}